refactor: move editor-specific code to separate module
JS & SCSS
