+++ src/editor/style.scss
@@ -2,14 +2,6 @@
 	max-width: 560px;
 }
 
-.editor-post-title {
-	display: none;
-}
-
-.components-button.editor-post-publish-button__button.is-primary {
-	display: none;
-}
-
 .wp-block-button {
 	margin-left: auto !important;
 	margin-right: auto !important;
@@ -150,4 +142,9 @@
 			margin-left: 12px;
 		}
 	}
+}
+
+.wp-block-buttons .block-editor-block-list__layout {
+	display: flex;
+	justify-content: space-between;
 }
