+++ src/newsletter-editor/sidebar/index.js
